+++ packages/client/src/components/App.css
@@ -1,0 +1,4 @@
+.Title {
+  text-align: center;
+  margin-top: 9em;
+}
