+++ packages/client/src/components/App.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import './App.css';
+
+export default function Title() {
+  return <h1 className="Title">Hi, Celina </h1>;
+}
